requirements.txt
pypdf
langchain
torch
accelerate
bitsandbytes
ctransformers
sentence_transformers
faiss_cpu
chainlit
huggingface_hub
os
numpy
random
tf-keras
tensorflow